Add files via upload
“Self-organizing maps (SOMs) are a data visualization technique invented by Professor Teuvo Kohonen which reduce the dimensions of data through the use of self-organizing neural networks.”  

	A SOM is an artificial neural network type where the learning is unsupervised and there are no hidden layers. Algorithm produces a direct map between the training set and the output network. SOMs visualize high dimensional data which humans cannot visualize and help humans understand the high dimensional data. 
The customer segmentation problem was chosen to be analyzed using the SOM. Fraud detection is a process that identify segments of customers and determine the segments will contain the customers that potantially cheated. 
Our aim is to predict the potential frauds and reply the customers credit card applications. The output for an application will be either approved (1) or not approved (0).